backend/requirements.txt
asgiref==3.11.1
black==26.1.0
celery==5.5.3
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
feedparser==6.0.12
idna==3.11
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
redis==6.4.0
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
uv==0.10.7
